+++ examples/pprobuf/checkstream.go
@@ -5,6 +5,8 @@ package main
 import (
 	"context"
 	"log"
+	"net/http"
+	_ "net/http/pprof"
 	"os"
 
 	_ "github.com/joho/godotenv/autoload"
@@ -13,24 +15,24 @@ import (
 )
 
 func main() {
+	go http.ListenAndServe(":6060", nil)
 	ctx := context.Background()
 	cfg := itdgo.Config{
 		RefreshToken: os.Getenv("REFRESH_TOKEN"),
-		UserAgent:    os.Getenv("USERAGENT"),
+		UserAgent:    os.Getenv("USER_AGENT"),
 	}
+
 	client, err := itdgo.New(ctx, cfg)
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	iter := client.Notifications.NewNotifications(10)
-
-	for iter.HasMore() {
-		notifications, err := iter.Next(ctx)
-		if err != nil {
+	stream, errs := client.Notifications.Stream(ctx)
+	for {
+		select {
+		case err := <-errs:
 			log.Fatal(err)
-		}
-		for _, notification := range notifications {
+		case notification := <-stream:
 			pp.Println(notification)
 		}
 	}