+++ examples/search/get_hashtag_feed.go
@@ -6,6 +6,7 @@ import (
 	"context"
 	"log"
 	"os"
+	"time"
 
 	_ "github.com/joho/godotenv/autoload"
 	"github.com/k0kubun/pp"
@@ -16,22 +17,25 @@ func main() {
 	ctx := context.Background()
 	cfg := itdgo.Config{
 		RefreshToken: os.Getenv("REFRESH_TOKEN"),
-		UserAgent:    os.Getenv("USERAGENT"),
+		UserAgent:    os.Getenv("USER_AGENT"),
+		RetryDelay:   4 * time.Second,
 	}
+
 	client, err := itdgo.New(ctx, cfg)
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	iter := client.Notifications.NewNotifications(10)
+	iter, err := client.Search.NewHashtagPosts("nowkie", 5)
+	if err != nil {
+		log.Fatal(err)
+	}
 
 	for iter.HasMore() {
-		notifications, err := iter.Next(ctx)
+		posts, err := iter.Next(ctx)
 		if err != nil {
 			log.Fatal(err)
 		}
-		for _, notification := range notifications {
-			pp.Println(notification)
-		}
+		pp.Println(posts)
 	}
 }